refactor(web): address PR #505 review feedback
- Scope the teaser-card tab guard to the tab header so clicks in the
  chart body bubble to the card link and navigate to /analysis again
  (Codex P2), with a regression test.
- Extract the duplicated callable-mock scaffolding from the four admin
  specs into a shared callable-functions.testing helper; *.testing.ts
  is excluded from the app tsconfig and typed via the spec tsconfig
  (CodeRabbit + Copilot).
