+++ web/src/app/admin/admin-page.component.spec.ts
@@ -9,32 +9,14 @@ import { of } from 'rxjs';
 import { AdminPageComponent } from './admin-page.component';
 import { AdminUser } from './admin-page.models';
 import { CallableFunctionsService } from './callable-functions.service';
+import {
+  CallableRecord,
+  createCallablesMock,
+} from './callable-functions.testing';
 import { DeleteUserDialogComponent } from './delete-user-dialog.component';
 import { UserDetailsDialogComponent } from './user-details-dialog.component';
 
-interface CallableRecord {
-  name: string;
-  impl: (data: unknown) => Promise<{ data: unknown }>;
-}
-
-// Faked via DI instead of vi.mock('@angular/fire/functions'): module
-// mocks of the fire package break whenever the spec bundler moves it
-// into a shared chunk (the mocked `Functions` token and the component's
-// bundled one are then different classes -> NG0201), and which admin
-// spec breaks shifts with the workspace's overall spec-file set.
-const callablesMock = { call: vi.fn() };
-
-function setupCallables(records: CallableRecord[]): void {
-  callablesMock.call.mockImplementation((name: string) => {
-    const match = records.find((r) => r.name === name);
-    if (!match) {
-      return async () => {
-        throw new Error(`Unexpected callable: ${name}`);
-      };
-    }
-    return match.impl;
-  });
-}
+const { callablesMock, setupCallables } = createCallablesMock();
 
 describe('AdminPageComponent', () => {
   let fixture: ComponentFixture<AdminPageComponent>;
