+++ web/src/app/stats/components/analysis-teaser-card/analysis-teaser-card.component.html
@@ -24,11 +24,8 @@
   </mat-card-header>
   <mat-card-content>
     @if (tabs().length > 0) {
-      <!--
-        Tab interaction must not bubble into the card's link-style
-        click/enter/space handlers — selecting an exercise tab is not
-        an intent to leave the dashboard.
-      -->
+      <!-- Guards only tab-header events; chart-body clicks still
+           activate the card link. See stopCardActivation. -->
       <div
         class="teaser-tabs"
         role="presentation"
