+++ web/tsconfig.app.json
@@ -8,6 +8,7 @@
   "exclude": [
     "src/**/*.spec.ts",
     "src/**/*.test.ts",
+    "src/**/*.testing.ts",
     "jest.config.ts",
     "jest.config.cts",
     "src/test-setup.ts"
