+++ web/tsconfig.spec.json
@@ -7,5 +7,10 @@
     "moduleResolution": "bundler"
   },
   "files": ["src/test-setup.ts"],
-  "include": ["src/**/*.test.ts", "src/**/*.spec.ts", "src/**/*.d.ts"]
+  "include": [
+    "src/**/*.test.ts",
+    "src/**/*.spec.ts",
+    "src/**/*.d.ts",
+    "src/**/*.testing.ts"
+  ]
 }
